requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.11.2

# Azure AI and Semantic Kernel
semantic-kernel==1.27.2
azure-ai-inference==1.0.0b9
azure-ai-projects==0.1.0
azure-core==1.33.0
azure-identity==1.14.0
azure-keyvault-secrets==4.9.0

# SharePoint access
Office365-REST-Python-Client==2.6.1
msal==1.32.0

# Data processing
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
matplotlib==3.7.2
seaborn==0.13.2

# Code interpreter dependencies
ipykernel==6.29.5
nbformat==5.9.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7

# Testing
pytest==7.4.2
pytest-mock==3.11.1

# Azure Storage
azure-storage-blob==12.17.0

# Web API server
flask==3.1.0
flask-cors==5.0.1
werkzeug==2.3.7
gunicorn==21.2.0

# For async operations
httpx==0.24.1
aiohttp==3.8.5

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1
tqdm==4.66.1
